Update SNEMI3D configuration and model classes for improved training performance
This commit modifies the `snemi3d_microns.yaml` configuration to increase `num_workers` from 12 to 16, enhancing data loading efficiency. The precision setting is updated from `32-true` to `bf16-mixed`, optimizing computational performance. Additionally, the dropout rate is adjusted from `0.0` to `0.1` in the model configuration, and the `benchmark` flag is set to `false` to improve training stability. Changes in the `cosmospredict3d_model.py` and `cosmostransfer3d_model.py` files include updates to the VAE loading process and adjustments to the temporal compression parameter, reducing it from `8` to `4`. These modifications aim to streamline the training process and enhance overall model performance.

+++ neurons/modules/cosmos.py
@@ -359,7 +359,7 @@ class BaseCosmosModule(pl.LightningModule):
 
         fg_mask = targets["labels"] > 0
         if fg_mask.any():
-            ins_pred, _, _ = self._clusterer(predictions["instance"], fg_mask)
+            ins_pred, _, _ = self._clusterer(predictions["instance"].float(), fg_mask)
             ins_gt = targets["labels"]
 
             ins_ari = compute_per_batch_ari(ins_pred, ins_gt)
